fix(checker): split cross-file heritage helpers
